Several fixes
